feat: more fixes
